Refactor main.py and basemodels.py
This commit refactors the main.py file by adding an import statement for the basemodels module.
